Restructured some files, renamed register folder to auth, renamed appcss to stylescss
