+++ wc-tracker/src/components/auth/register.js
